requirements.txt
Pillow
PyAutoGUI
pynput
opencv-python
pywin32
NumPy
matplotlib